Fix acceptance tests that depend on a processing state
That kind of tests are highly dependent on the machine running
the tests. That caused Travis to fail too frequently.
Now, instead of checking for a specific state, we check
if a state is valid or if we can retrieve a Processing.

Signed off by: Heitor Reis <marcheing@gmail.com>